+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Fixed: `--ignore-path` and `--report-needless-disables` no longer fails when used together.
 -   Fixed: the `indentation` rule now correctly handles `_` hacks on property names.
+-   Added: `ignore: ["after-declaration"]` option to `declaration-empty-line-before`.
 
 # 7.2.0